Add controller option to force a full reload.
A controller can now call #sans_turbolinks! to ensure its response will be rejected by the client and they will perform a full reload instead.

There are plenty of ways this might be optimised, such as the controller not rendering the page when the initial turbolinks request happens (then the client would send something back with the full request triggering the controller to do a real render), but I'm not sure of the best way yet.