Improve full reload from controller by rendering nothing initially.
This extends the previous function so that the initial turbolinks response is blank (rather than being a full rendered page, often requiring a lot of extra SQL queries too) and then the full page is only rendered with a full reload. This works by first marking all turbolinks requests with a header and then detecting that header on the serverside and rendering nothing (as well as the previous "X-Sans-Turbolinks" header that will make the client do a full reload) in response.

The new option for this is `render_sans_turbolinks`, which should replace a `render` call within that controller action. If the request header is detected it will finish with `render nothing: true`, otherwise if it receives any params it will pass them to a `render` call. If no params are received it'll return to the controller having only set response headers indicating it's a sans-turbolinks request and will carry on as usual (often this means hitting the end of the user-defined controller action and having rails automatically render the appropriate view).

The old option of just rendering the response header has been kept in as one may want to use it on its own.